feat: better encoding go routines and progress notification, support to process custom thumbnails, and better general queue naming

+++ metadata/src/main/java/com/shin/metadata/dto/ContentDetails.java
@@ -4,7 +4,7 @@ import com.shin.metadata.model.enums.VideoLanguage;
 
 public record ContentDetails(
         String resolutions,
-        Double duration,
+        Long duration,
         String processedPath,
         VideoLanguage defaultLanguage,
         String publishedLocale,
